+++ packages/context/src/lib/context-manager/context-list/context-list-binding.directive.ts
@@ -8,14 +8,16 @@ import {
 import { Subscription } from 'rxjs';
 import { debounceTime } from 'rxjs/operators';
 
-import { MessageService, LanguageService } from '@igo2/core';
+import { MessageService, LanguageService, StorageService } from '@igo2/core';
+import { AuthService } from '@igo2/auth';
 import { ConfirmDialogService } from '@igo2/common';
 import { MapService } from '@igo2/geo';
 
 import {
   Context,
   DetailedContext,
-  ContextsList
+  ContextsList,
+  ContextUserPermission
 } from '../shared/context.interface';
 import { ContextService } from '../shared/context.service';
 import { ContextListComponent } from './context-list.component';
@@ -44,14 +46,7 @@ export class ContextListBindingDirective implements OnInit, OnDestroy {
     const map = this.mapService.getMap();
     const contextFromMap = this.contextService.getContextFromMap(map);
 
-    const changes: any = {
-      layers: contextFromMap.layers,
-      map: {
-        view: contextFromMap.map.view
-      }
-    };
-
-    this.contextService.update(context.id, changes).subscribe(() => {
+    const msgSuccess = () => {
       const translate = this.languageService.translate;
       const message = translate.instant(
         'igo.context.contextManager.dialog.saveMsg',
@@ -63,6 +58,27 @@ export class ContextListBindingDirective implements OnInit, OnDestroy {
         'igo.context.contextManager.dialog.saveTitle'
       );
       this.messageService.success(message, title);
+    };
+
+    if (context.imported) {
+      contextFromMap.title = context.title;
+      this.contextService.delete(context.id, true);
+      this.contextService.create(contextFromMap).subscribe((contextCreated) => {
+        this.contextService.loadContext(contextCreated.uri);
+        msgSuccess();
+      });
+      return;
+    }
+
+    const changes: any = {
+      layers: contextFromMap.layers,
+      map: {
+        view: contextFromMap.map.view
+      }
+    };
+
+    this.contextService.update(context.id, changes).subscribe(() => {
+      msgSuccess();
     });
   }
 
@@ -101,20 +117,22 @@ export class ContextListBindingDirective implements OnInit, OnDestroy {
       .open(
         translate.instant('igo.context.contextManager.dialog.confirmDelete')
       )
-      .subscribe(confirm => {
+      .subscribe((confirm) => {
         if (confirm) {
-          this.contextService.delete(context.id).subscribe(() => {
-            const message = translate.instant(
-              'igo.context.contextManager.dialog.deleteMsg',
-              {
-                value: context.title
-              }
-            );
-            const title = translate.instant(
-              'igo.context.contextManager.dialog.deleteTitle'
-            );
-            this.messageService.info(message, title);
-          });
+          this.contextService
+            .delete(context.id, context.imported)
+            .subscribe(() => {
+              const message = translate.instant(
+                'igo.context.contextManager.dialog.deleteMsg',
+                {
+                  value: context.title
+                }
+              );
+              const title = translate.instant(
+                'igo.context.contextManager.dialog.deleteTitle'
+              );
+              this.messageService.info(message, title);
+            });
         }
       });
   }
@@ -140,13 +158,69 @@ export class ContextListBindingDirective implements OnInit, OnDestroy {
     });
   }
 
+  @HostListener('create', ['$event'])
+  onCreate(opts: { title: string; empty: boolean }) {
+    const { title, empty } = opts;
+    const context = this.contextService.getContextFromMap(
+      this.component.map,
+      empty
+    );
+    context.title = title;
+    this.contextService.create(context).subscribe(() => {
+      const translate = this.languageService.translate;
+      const titleD = translate.instant(
+        'igo.context.bookmarkButton.dialog.createTitle'
+      );
+      const message = translate.instant(
+        'igo.context.bookmarkButton.dialog.createMsg',
+        {
+          value: context.title
+        }
+      );
+      this.messageService.success(message, titleD);
+      this.contextService.loadContext(context.uri);
+    });
+  }
+
+  @HostListener('filterPermissionsChanged')
+  loadContexts() {
+    const permissions = ['none'];
+    for (const p of this.component.permissions) {
+      if (p.checked === true || p.indeterminate === true) {
+        permissions.push(p.name);
+      }
+    }
+    this.component.showHidden
+      ? this.contextService.loadContexts(permissions, true)
+      : this.contextService.loadContexts(permissions, false);
+  }
+
+  @HostListener('showHiddenContexts')
+  showHiddenContexts() {
+    this.component.showHidden = !this.component.showHidden;
+    this.storageService.set('contexts.showHidden', this.component.showHidden);
+    this.loadContexts();
+  }
+
+  @HostListener('show', ['$event'])
+  onShowContext(context: DetailedContext) {
+    this.contextService.showContext(context.id).subscribe();
+  }
+
+  @HostListener('hide', ['$event'])
+  onHideContext(context: DetailedContext) {
+    this.contextService.hideContext(context.id).subscribe();
+  }
+
   constructor(
     @Self() component: ContextListComponent,
     private contextService: ContextService,
     private mapService: MapService,
     private languageService: LanguageService,
     private confirmDialogService: ConfirmDialogService,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private auth: AuthService,
+    private storageService: StorageService
   ) {
     this.component = component;
   }
@@ -154,13 +228,16 @@ export class ContextListBindingDirective implements OnInit, OnDestroy {
   ngOnInit() {
     // Override input contexts
     this.component.contexts = { ours: [] };
+    this.component.showHidden = this.storageService.get(
+      'contexts.showHidden'
+    ) as boolean;
 
-    this.contexts$$ = this.contextService.contexts$.subscribe(contexts =>
+    this.contexts$$ = this.contextService.contexts$.subscribe((contexts) =>
       this.handleContextsChange(contexts)
     );
 
     this.defaultContextId$$ = this.contextService.defaultContextId$.subscribe(
-      id => {
+      (id) => {
         this.component.defaultContextId = id;
       }
     );
@@ -168,9 +245,39 @@ export class ContextListBindingDirective implements OnInit, OnDestroy {
     // See feature-list.component for an explanation about the debounce time
     this.selectedContext$$ = this.contextService.context$
       .pipe(debounceTime(100))
-      .subscribe(context => (this.component.selectedContext = context));
+      .subscribe((context) => (this.component.selectedContext = context));
 
-    this.contextService.loadContexts();
+    this.auth.authenticate$.subscribe((authenticate) => {
+      if (authenticate) {
+        this.contextService.getProfilByUser().subscribe((profils) => {
+          this.component.users = profils;
+          this.component.permissions = [];
+          const profilsAcc = this.component.users.reduce((acc, cur) => {
+            acc = acc.concat(cur);
+            acc = cur.childs ? acc.concat(cur.childs) : acc;
+            return acc;
+          }, []);
+          for (const user of profilsAcc) {
+            const permission: ContextUserPermission = {
+              name: user.name,
+              checked: this.storageService.get(
+                'contexts.permissions.' + user.name
+              ) as boolean
+            };
+            this.component.permissions.push(permission);
+          }
+          const permissions = ['none'];
+          for (const p of this.component.permissions) {
+            if (p.checked === true || p.indeterminate === true) {
+              permissions.push(p.name);
+            }
+          }
+          this.component.showHidden
+            ? this.contextService.loadContexts(permissions, true)
+            : this.contextService.loadContexts(permissions, false);
+        });
+      }
+    });
   }
 
   ngOnDestroy() {